Move duplicate code to test helper file
Re #23027
