+++ qt/scientific_interfaces/test/ISISReflectometry/Instrument/InstrumentPresenterTest.h
@@ -8,6 +8,7 @@
 #define MANTID_CUSTOMINTERFACES_INSTRUMENTPRESENTERTEST_H_
 
 #include "../../../ISISReflectometry/GUI/Instrument/InstrumentPresenter.h"
+#include "../ModelCreationHelpers.h"
 #include "../ReflMockObjects.h"
 #include "MantidAPI/FrameworkManager.h"
 #include "MantidAPI/MatrixWorkspace.h"
@@ -253,7 +254,7 @@ public:
   }
 
   void testRestoreDefaultsNotifiesMainPresenter() {
-    auto defaultOptions = expectDefaults(makeModel());
+    auto defaultOptions = expectDefaults(makeEmptyInstrument());
     auto presenter = makePresenter(std::move(defaultOptions));
     EXPECT_CALL(m_mainPresenter, notifyRestoreDefaultsRequested())
         .Times(AtLeast(1));
@@ -339,16 +340,6 @@ private:
   NiceMock<MockInstrumentView> m_view;
   NiceMock<MockBatchPresenter> m_mainPresenter;
 
-  Instrument makeModel() {
-    auto wavelengthRange = RangeInLambda(0.0, 0.0);
-    auto monitorCorrections = MonitorCorrections(
-        0, false, RangeInLambda(0.0, 0.0), RangeInLambda(0.0, 0.0));
-    auto detectorCorrections =
-        DetectorCorrections(false, DetectorCorrectionType::VerticalShift);
-    return Instrument(std::move(wavelengthRange), std::move(monitorCorrections),
-                      std::move(detectorCorrections));
-  }
-
   Instrument
   makeModelWithMonitorOptions(MonitorCorrections monitorCorrections) {
     auto wavelengthRange = RangeInLambda(0.0, 0.0);
@@ -379,8 +370,8 @@ private:
   InstrumentPresenter
   makePresenter(std::unique_ptr<IInstrumentOptionDefaults> defaultOptions =
                     std::make_unique<MockInstrumentOptionDefaults>()) {
-    auto presenter =
-        InstrumentPresenter(&m_view, makeModel(), std::move(defaultOptions));
+    auto presenter = InstrumentPresenter(&m_view, makeEmptyInstrument(),
+                                         std::move(defaultOptions));
     presenter.acceptMainPresenter(&m_mainPresenter);
     return presenter;
   }
