+++ qt/scientific_interfaces/test/ISISReflectometry/Runs/RunsPresenterTest.h
@@ -9,6 +9,7 @@
 
 #include "../../../ISISReflectometry/GUI/Runs/RunsPresenter.h"
 #include "../../../ISISReflectometry/Reduction/RunsTable.h"
+#include "../ModelCreationHelpers.h"
 #include "../ReflMockObjects.h"
 #include "../RunsTable/MockRunsTablePresenter.h"
 #include "../RunsTable/MockRunsTableView.h"
@@ -340,17 +341,8 @@ private:
     return presenter;
   }
 
-  Row makeRowWithRun(std::string const &run) {
-    return Row(std::vector<std::string>{"12345"}, 0.5, TransmissionRunPair(),
-               RangeInQ(), boost::none, ReductionOptionsMap(),
-               ReductionWorkspaces(std::vector<std::string>{run},
-                                   TransmissionRunPair()));
-  }
-
   RunsTable makeRunsTableWithContent() {
-    auto rows = std::vector<boost::optional<Row>>{makeRowWithRun("12345")};
-    auto groups = std::vector<Group>{Group("Group 1", rows)};
-    auto reductionJobs = ReductionJobs(groups);
+    auto reductionJobs = makeReductionJobsWithSingleRowGroup();
     return RunsTable(m_instruments, m_thetaTolerance, reductionJobs);
   }
